chatapp/requirements.txt
langchain==0.1.5
langchain-community==0.0.17
langchain-openai==0.0.5
boto3==1.34.37
python-dotenv==1.0.1
streamlit==1.31.1
langchain_aws==0.1.2